+++ src/Logging/Traits/AppNameTrait.php
@@ -4,6 +4,9 @@ declare(strict_types = 1);
 
 namespace AvtoDev\EventsLogLaravel\Logging\Traits;
 
+use Illuminate\Container\Container;
+use Illuminate\Contracts\Config\Repository;
+
 trait AppNameTrait
 {
     /**
@@ -14,11 +17,9 @@ trait AppNameTrait
     protected function getAppName(): ?string
     {
         try {
-            return config()->get('app.name');
+            return Container::getInstance()->make(Repository::class)->get('app.name');
         } catch (\Throwable $e) {
-            //
+            return null;
         }
-
-        return null;
     }
 }